# openapi.yaml
openapi: 3.1.0
info:
  title: ForgeConfig API
  description: ForgeConfig — REST API for managing network device provisioning, configuration, backups, IPAM, and more.
  version: 0.2.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/docker/containers:
    get:
      tags: [Docker]
      summary: List managed containers
      responses:
        "200":
          description: Array of containers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestContainer"
    post:
      tags: [Docker]
      summary: Spawn a test container
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpawnContainerRequest"
      responses:
        "201":
          description: Spawned container
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestContainer"
  /api/docker/containers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      tags: [Docker]
      summary: Remove container
      responses:
        "200":
          $ref: "#/components/responses/Message"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Message:
      description: Success message
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
...
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    TestContainer:
      type: object
      required: [id, name, hostname, mac, ip, status, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        hostname:
          type: string
        mac:
          type: string
        ip:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    SpawnContainerRequest:
      type: object
      properties:
        hostname:
          type: string
        mac:
          type: string
        vendor_class:
          type: string
          description: DHCP Option 60 vendor class identifier
        config_method:
          type: string
          enum: [tftp, http, both]
        image:
          type: string
          description: Docker image to use
        topology_id:
          type: integer
          format: int64
        topology_role:
          type: string
...